% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{AbaloneIdt}
\alias{AbaloneIdt}
\title{AbaloneIdt data example}
\format{
An object of class \code{data.frame} (inherits from \code{symbolic_tbl}) with 24 rows and 7 columns.
}
\usage{
data(AbaloneIdt)
}
\description{
AbaloneIdt interval data example.
}
\examples{
data(AbaloneIdt)
ggInterval_index(AbaloneIdt, aes(x = Length))
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
