% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{BLOOD}
\alias{BLOOD}
\title{BLOOD data example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}, \code{symbolic_tbl}) with 14 rows and 3 columns.
}
\usage{
data(BLOOD)
}
\description{
BLOOD interval data example.
}
\examples{
data(BLOOD)
ggInterval_minmax(BLOOD, aes(x = Hematocrit))
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
