% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Cardiological}
\alias{Cardiological}
\title{Cardiological data example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11 rows and 3 columns.
}
\source{
\url{https://CRAN.R-project.org/package=RSDA}
}
\usage{
data(Cardiological)
}
\description{
Cardiological interval data example.
}
\examples{
data(Cardiological)
ggInterval_index(Cardiological, aes(x = Syst))
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
