% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Environment}
\alias{Environment}
\title{Environment data example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 17 columns.
}
\usage{
data(Environment)
}
\description{
Environment interval and modal data example.
}
\examples{
data(Environment)
ggInterval_radar(Environment,
 plotPartial = 2,
 showLegend = FALSE,
 base_circle = TRUE,
 base_lty = 2,
 addText = FALSE)
}
\keyword{datasets}
