% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggInterval_3Dscatter.R
\name{ggInterval_3Dscatter}
\alias{ggInterval_3Dscatter}
\title{3D scatter plot for interval data}
\usage{
ggInterval_3Dscatter(data = NULL,mapping = aes(NULL),scale=FALSE)
}
\arguments{
\item{data}{A ggSDA object. It can also be either RSDA object or
classical data frame, which will be automatically convert to ggSDA
data.}

\item{mapping}{Set of aesthetic mappings created by aes() or aes_().
If specified and inherit. aes = TRUE (the default),
it is combined with the default mapping at the top level of
the plot. You must supply mapping if there is no plot mapping.
It is the same as the mapping of ggplot2.}

\item{scale}{A boolean variable,TRUE, standardlize data. FALSE, not standardlize.
If variance is too large(or small) or the difference between
two variables are too large,it will be distortion or unseeable,which may
happen in different units or others. So, a standardlize way is necessary.}
}
\value{
Return a ggplot2 object (It will still be 2-Dimension).
}
\description{
Visualize the three continuous variable distribution
by collecting all vertices in each interval to form a shape of
cube.Also show the difference between each group.
}
\examples{
ggInterval_3Dscatter(facedata[1:5, ], aes(x = BC, y = EH, z = GH))

}
