% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{iris.i}
\alias{iris.i}
\title{iris.i data example}
\format{
An object of class \code{data.frame} (inherits from \code{symbolic_tbl}) with 3 rows and 4 columns.
}
\usage{
data(iris.i)
}
\description{
iris.i interval data example.
}
\examples{
data(iris.i)
ggInterval_index(iris.i, aes(x = Sepal.Length))
}
\keyword{datasets}
