% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mtcars.i}
\alias{mtcars.i}
\title{mtcars.i data example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}, \code{symbolic_tbl}) with 5 rows and 11 columns.
}
\usage{
data(mtcars.i)
}
\description{
mtcars.i interval and modal data example.
}
\examples{
data(mtcars.i)
ggInterval_index(mtcars.i, aes(x = mpg))
}
\keyword{datasets}
