% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{oils}
\alias{oils}
\title{oils data example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 4 columns.
}
\usage{
data(oils)
}
\description{
oils interval data example.
}
\examples{
data(oils)
ggInterval_scatter(oils, aes(x = GRA, y = IOD))
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
