% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_sym.r
\name{scale_sym}
\alias{scale_sym}
\title{scale for symbolic data table}
\usage{
scale.sym(data = NULL)
}
\arguments{
\item{data}{A ggESDA object. It can also be either RSDA object or
classical data frame, which will be automatically convert to ggESDA
data.}
}
\value{
Return a scale ggESDA object.
}
\description{
scale for symbolic data table
}
\examples{

#For all interval-valued
scale.sym(facedata)

#For both interval-valued and modal multi-valued
scale.sym(mtcars.i)

}
