% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.default}
\alias{summary.symbolic_tbl}
\alias{summary.symbolic_interval}
\alias{summary.symbolic_modal}
\title{summary for symbolic data table}
\usage{
summary(object, ...)

\method{summary}{default}(object, ...)

\method{summary}{symbolic_tbl}(x, ...)

\method{summary}{symbolic_interval}(x, ...)

\method{summary}{symbolic_modal}(x, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}

\item{x}{a result of the default method of summary().}

\item{maxsum}{integer, indicating how many levels should be shown for factors.}

\item{digits}{integer, used for number formatting with signif() (for summary.default) or format() (for summary.data.frame). In summary.default, if not specified (i.e., missing(.)), signif() will not be called anymore (since R >= 3.4.0, where the default has been changed to only round in the print and format methods).}

\item{quantile.type}{integer code used in quantile(*, type=quantile.type) for the default method.}
}
\value{
Return a summary table.
}
\description{
summary for symbolic data table
}
\examples{

#For all interval-valued
summary(facedata)

#For both interval-valued and modal multi-valued
summary(Environment)


}
\keyword{Symbolic}
\keyword{summary}
