% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.default}
\alias{summary.symbolic_tbl}
\alias{summary.symbolic_interval}
\alias{summary.symbolic_modal}
\title{summary for symbolic data table}
\usage{
summary(object, ...)

\method{summary}{default}(object, ...)

\method{summary}{symbolic_tbl}(object, ...)

\method{summary}{symbolic_interval}(object, ...)

\method{summary}{symbolic_modal}(object, summary_fun = "mean", ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}

\item{summary_fun}{only works when the symbolic_modal class input, it determine which summary function be applied for each modal.}
}
\value{
Return a summary table.
}
\description{
summary for symbolic data table
}
\examples{

#For all interval-valued
summary(facedata)

#For both interval-valued and modal multi-valued
summary(Environment)

summary(Environment$URBANICITY, summary_fun = "quantile")


}
\keyword{Symbolic}
\keyword{summary}
